requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
python-jose[cryptography]
passlib[bcrypt]
pydantic
python-dotenv
google-genai
email-validator
argon2-cffi
